requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
requests==2.25.1
redis==3.5.3
python-dotenv==0.17.0
